Added full VS2010 support.  There must be no problems with project deployment in other locations(all paths are relative).

+++ RushEngine/Eventing.h
@@ -20,15 +20,15 @@ public:
 		localtime_s(&TimeNow, &TmNow);
 		string DataToWrite = "\n";
 		DataToWrite += "Date : ";
-		DataToWrite += to_string(TimeNow.tm_mday);
+		DataToWrite += to_string(static_cast<long long>(TimeNow.tm_mday));
 		DataToWrite += "/";
-		DataToWrite += to_string(TimeNow.tm_mon + 1);
+		DataToWrite += to_string(static_cast<long long>(TimeNow.tm_mon + 1));
 		DataToWrite += "/";
-		DataToWrite += to_string(TimeNow.tm_year + 1900);
+		DataToWrite += to_string(static_cast<long long>(TimeNow.tm_year + 1900));
 		DataToWrite += " ";
-		DataToWrite += to_string(TimeNow.tm_hour);
+		DataToWrite += to_string(static_cast<long long>(TimeNow.tm_hour));
 		DataToWrite += ":";
-		DataToWrite += to_string(TimeNow.tm_min);
+		DataToWrite += to_string(static_cast<long long>(TimeNow.tm_min));
 		DataToWrite += "\n";
 		DataToWrite+= "Data Type : ";
 		DataToWrite += Type;
@@ -40,10 +40,12 @@ public:
 		ofstream Log("logs.txt", std::ios_base::app | std::ios_base::out);
 		Log << DataToWrite;
 		Log.close();
+		/*
 		if (Throw == true)
 		{
 			throw Data;
 		}
+		*/
 	}
 	void GraphicsError(string Error)
 	{
